Move lazy lambda tests into their own file
